fix join meeting and other refactoring
